Add get_error_message & connect_mqtt funcs
